+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Adding `xan search -u/--url-prefix`.
 * Adding `xan network -f nodelist` and `xan network --degrees`.
 * Adding various format aliases to `xan from`.
+* Adding `xan explode -D/--drop-empty`.
 
 *Fixes*